+++ ikii-back-microservices/ikii-back-business/src/main/java/mx/ikii/business/repository/IBusinessRepository.java
@@ -1,5 +1,8 @@
 package mx.ikii.business.repository;
 
+import java.util.List;
+
+import org.bson.types.ObjectId;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
@@ -9,5 +12,9 @@ import mx.ikii.commons.persistence.collection.Business;
 public interface IBusinessRepository extends MongoRepository<Business, String> {
 	
 	Business findByName(String name);
+	
+	List<Business> findByCustomerId(ObjectId customerId);
+	
+
 
 }