+++ ikii-back-microservices/ikii-back-business/src/main/java/mx/ikii/business/service/IBusinessService.java
@@ -23,4 +23,6 @@ public interface IBusinessService {
 
 	void delete(String id);
 
+	Business findByCustomerId(String customerId);
+
 }