+++ ikii-back-microservices/ikii-back-business/src/main/java/mx/ikii/business/service/IBusinessServiceWrapper.java
@@ -25,4 +25,6 @@ public interface IBusinessServiceWrapper {
 
 	void delete(String id);
 
+	BusinessResponse getByCustomerId(String customerId);
+
 }